Rename & Move Deno LSP types
