+++ netlify/edge-functions/api-pkg.ts
@@ -2,7 +2,7 @@ import { compareEtag } from 'https://deno.land/std@0.151.0/http/util.ts';
 import { calculate as getETag } from 'https://deno.land/x/oak@v10.6.0/etag.ts';
 
 import type { EdgeFunction } from 'netlify:edge';
-import type { CompletionList } from './api/types.ts';
+import type { CompletionList } from './lib/deno-types.ts';
 
 const pattern = new URLPattern({
 	pathname: '/api/x/:package([a-zA-Z0-9_-]*)',
